separate stream rest api integration tests from websocket tests
