+++ test/integration/StreamrClient.test.js
@@ -12,7 +12,6 @@ describe('StreamrClient', () => {
     const name = `StreamrClient-integration-${Date.now()}`
 
     let client
-    let createdStream
 
     const createClient = (opts = {}) => new StreamrClient({
         url: config.websocketUrl,
@@ -49,46 +48,18 @@ describe('StreamrClient', () => {
         }
     })
 
-    describe('Stream creation', () => {
-        it('createStream', () => client.createStream({
-            name,
-        })
-            .then((stream) => {
+    describe('Pub/Sub', () => {
+        let createdStream
+
+        beforeAll(() => {
+            assert(client.isConnected())
+            return client.createStream({
+                name,
+            }).then((stream) => {
                 createdStream = stream
                 assert(stream.id)
                 assert.equal(stream.name, name)
-            }))
-
-        it('getOrCreate an existing Stream', () => client.getOrCreateStream({
-            name,
-        })
-            .then((existingStream) => {
-                assert.equal(existingStream.id, createdStream.id)
-                assert.equal(existingStream.name, createdStream.name)
-            }))
-
-        it('getOrCreate a new Stream', () => {
-            const newName = Date.now()
-                .toString()
-            return client.getOrCreateStream({
-                name: newName,
             })
-                .then((newStream) => {
-                    assert.notEqual(newStream.id, createdStream.id)
-                })
-        })
-    })
-
-    describe('Stream.update', () => {
-        it('can change stream name', () => {
-            createdStream.name = 'New name'
-            return createdStream.update()
-        })
-    })
-
-    describe('Data production', () => {
-        beforeAll(() => {
-            assert(client.isConnected())
         })
 
         it('Stream.produce', () => createdStream.produce({
@@ -142,44 +113,5 @@ describe('StreamrClient', () => {
                 })
             })
         })
-    })
-
-    describe('Stream configuration', () => {
-        it('Stream.detectFields', (done) => {
-            client.produceToStream(createdStream.id, {
-                foo: 'bar',
-                count: 0,
-            })
-
-            // Need time to propagate to storage
-            setTimeout(() => {
-                createdStream.detectFields().then((stream) => {
-                    assert.deepEqual(
-                        stream.config.fields,
-                        [
-                            {
-                                name: 'foo',
-                                type: 'string',
-                            },
-                            {
-                                name: 'count',
-                                type: 'number',
-                            },
-                        ],
-                    )
-                    done()
-                })
-            }, 5000)
-        }, 10000)
-    })
-
-    describe('Stream permissions', () => {
-        it('Stream.getPermissions', () => createdStream.getPermissions().then((permissions) => {
-            assert.equal(permissions.length, 3) // read, write, share for the owner
-        }))
-    })
-
-    describe('Stream deletion', () => {
-        it('Stream.delete', () => createdStream.delete())
     })
 })
